feat: create booking endpoint, fix time block model for population, and fix discord errors that the changes caused
