+++ src/models/timeBlock.model.ts
@@ -1,10 +1,9 @@
-import { Schema, model, Types } from 'mongoose';
-import booking from './booking.model';
+import { Schema, model, Types, Document, PopulatedDoc } from 'mongoose';
 import { Booking } from './booking.model'
 
 interface TimeBlock {
     // Users that are signed up for time block
-    users: [Booking];
+    bookings: PopulatedDoc<Booking & Document>;
     // Id of the room section that the time block corresponds to
     sectionId: Types.ObjectId;
     startsAt: Date;
@@ -13,7 +12,7 @@ interface TimeBlock {
 
 const timeBlockSchema = new Schema<TimeBlock>({
     // Users that are signed up for time block
-    users: [booking],
+    bookings: [{ type: Types.ObjectId, ref: 'Booking' }],
     // Id of the room section that the time block corresponds to
     sectionId: Types.ObjectId,
     startsAt: Date,
