--- src/routes/api/bookings/editBooking.ts
@@ -1,7 +1,0 @@
-import { Request, Response } from 'express';
-
-const editBookingRoute = async (req: Request, res: Response): Promise<void> => {
-    
-};
-
-export default editBookingRoute;
