+++ src/routes/api/index.ts
@@ -3,6 +3,7 @@ import userRoutes from './users';
 import roomRoutes from './rooms';
 import issueRoutes from './issues';
 import sectionRoutes from './sections';
+import bookingRoutes from './bookings';
 
 const router = Router();
 
@@ -10,6 +11,7 @@ router.use('/users', userRoutes);
 router.use('/rooms', roomRoutes);
 router.use('/issues', issueRoutes);
 router.use('/sections', sectionRoutes);
+router.use('/bookings', bookingRoutes);
 
 router.get('/', (req, res) => {
     res.send('Hello world!');
